Reorganize Valgrind wrapper to IMO make things easier to understand
